Updating the plotting and adding some functionality
The plots examining the GPs outputs have been improved to represent the correct support of the various types of data and comparisons to the 'truth'. Additional functionality is being added for the qm procedure to correctly propagate uncertainty and for the bg_lima script to work with all grid cells as an option. There are still some to-dos for these updates, which are included in the readme